add query api
