+++ requirements.txt
@@ -1,0 +1,6 @@
+joblib==1.1.0
+pandas==1.3.3
+spacy==3.1.3
+scikit-learn==1.0.2
+python-dateutil==2.8.2
+python-Levenshtein==0.12.2
