+++ src/main/python/search_engine/preprocess.py
@@ -4,8 +4,23 @@ import json
 import spacy
 from spacy.matcher import Matcher
 from dateutil.parser import parse
+import os
 
+original_dir = os.getcwd()
+def find_search_engine_path():
+    # Lấy đường dẫn tuyệt đối của thư mục hiện tại
+    current_directory = os.getcwd()
 
+    # Duyệt qua tất cả các thư mục và tệp tin trong cây thư mục bắt đầu từ thư mục hiện tại
+    for dirpath, dirnames, filenames in os.walk(current_directory):
+        # Kiểm tra xem thư mục hiện tại có tên là "search_engine" không
+        if "search_engine" in dirnames:
+            return os.path.join(dirpath, "search_engine")
+        
+    # Trả về None nếu không tìm thấy thư mục "search_engine" trong cây thư mục
+    return None
+
+os.chdir(find_search_engine_path())
 
 nlp = spacy.load("en_core_web_sm")
 matcher = Matcher(nlp.vocab)
@@ -19,6 +34,8 @@ with open('../../resources/assets/stopwords.txt') as file:
 
 with open('../../resources/assets/lemmatization-en.json') as file:
     lemmatizer = json.load(file)
+
+os.chdir(original_dir)
 
 class Preprocessor():
